Bug fixes: stringsdict, swipe delete, skip button in alert when adding new article, router animation and sort item's list in listsView.

+++ Listers/DataManaging/Persistence/Persistence.swift
@@ -172,7 +172,7 @@ struct PersistenceController {
                 newItem.id = UUID()
                 newItem.name = "Item \(itemNumber)"
                 newItem.notes = "This is item \(itemNumber)."
-                newItem.quantity = Double.random(in: 0...10)
+                newItem.quantity = Int16.random(in: 0...10)
                 newItem.creationDate = Date.now
                 newItem.favorite = Bool.random()
                 newItem.completed = Bool.random()
